phase3: split crown/global functions into src/crown.cpp
- Move QUAD, CalcVcmaxm, CalcRdark, and all crown geometry helpers
  (GetRadiusSlope, GetCrownIntarea, GetDensities*, LAI2dens*, UpdateLAI3D,
  UpdateCHM*, OutputCrownSliced, GetPPFDabove, GetCanopyEnvironment,
  AddCrownVolumeLayer) out of troll.cpp into src/crown.cpp
- Move CircleAreaUpdateCrownStatistic_template and
  LoopLayerUpdateCrownStatistic_template bodies into troll.hpp so they
  remain visible at all instantiation sites across future TUs
- Replace struct Context forward declaration in troll.hpp with
  #include "context.hpp" so template bodies can access ctx fields
