+++ src/troll.cpp
@@ -3382,469 +3382,6 @@ float Tree::StartTracking()
 }
 #endif
 
-////////////////////////////////////////////////
-/////// GLOBAL FUNCTIONS ///////////////////////
-////////////////////////////////////////////////
-
-#ifdef G0
-
-// Solves the quadratic equation. ROOT: finds smaller (-1), larger (1) root.
-float QUAD(float A, float B, float C, int ROOT)
-{
-    float quad = 0.0;
-    float discriminant = B * B - 4.0 * A * C;
-    // ROOT=-1: smaller root; ROOT=1: larger root
-    if (discriminant < 0)
-    {
-        cerr << "IMAGINARY ROOTS IN QUADRATIC" << endl;
-        quad = 0.0;
-    }
-    if (A == 0.0)
-    {
-        if (B == 0.0)
-            quad = 0.0;
-        else
-            quad = -C / B;
-    }
-    else
-        quad = (-B + ROOT * sqrt(discriminant)) / (2.0 * A);
-    return quad;
-}
-
-#endif
-
-//! Global functions to compute Vcmax (on a mass basis) and leaf dark respiration (on a leaf area basis)
-// newIM : to avoid repeating same functions in Tree and Species class. Was needed for the Species class to correctly initialise Rdark and compute LCP
-
-float CalcVcmaxm(float LMA, float Nmass, float Pmass)
-{
-    float SLA = 10000.0 / LMA;                                                                                                                               // in cm2 g-1
-    float Vcmaxm = pow(10.0, fminf((-1.56 + 0.43 * log10(Nmass * 1000.0) + 0.37 * log10(SLA)), (-0.80 + 0.45 * log10(Pmass * 1000.0) + 0.25 * log10(SLA)))); // this is equation 2 in Domingues et al 2010 PCE (coefficients from fig7) which made better fits than equation 1 (without LMA). Nmass and Pmass are given in g g-1, but should be in mg g-1 in equ 2 in Domingues et al. 2010, hence the mutiplication by 1000.
-    return (Vcmaxm);                                                                                                                                         //  in micromol C g-1 s-1
-}
-float CalcRdark(float LMA, float Nmass, float Pmass, float Vcmax)
-{
-    float Parea = Pmass * LMA;                                                                         // in g ctx.params.m-2
-    float Narea = Nmass * LMA;                                                                         // in g ctx.params.m-2
-    float Rdark = (1.3893 + (0.0728 * Narea) + (0.0015 * Parea) + (0.0095 * Vcmax) - (0.0358 * 26.2)); // in micromolC ctx.params.m-2 s-1 //since v.2.5: correction of Atkin et al. 2015 New phytologist formula. The original formula was based on mean-centered values (cf. Atkin et al. 2015, and the clarification/correction published afterwards), unfortunately only one formula (PFT-specific formula) with absolute values is provided and this is the one used here, cf. corrigendum TableS4): https://nph.onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1111%2Fnph.13253&file=nph13253-sup-0001-SupInfo.pdf Warning: Vcmax should be provided on an area basis.
-    return (Rdark);
-}
-
-#ifdef CROWN_UMBRELLA
-//! - Global function: the following function is not a member function of the Tree class object, but operates at the tree level and could be converted to member functions. They're therefore defined next to the other tree level functions. Whether conversion to Tree members adds any benefit in terms of performance should be tested.
-//! - the main function is a template to loop through one crown shell (i.e. a 1m ctx.diag.layer of the tree) and do something within the tree's canopy, such as allocating leaves or computing the flux. It is supported by a second template that simulates the actual loop. The crown shell can be bent via a function to simulate the umbrella like crown shapes
-//! - variables to be provided to the template:
-//! -# crown properties, including the crown position (row, col), the tree height, the crown radius, the crown depth, the fraction of filled voxels (inverse of gap fraction) as well as the ctx.diag.layer/shell counting from the tree top
-//! -# a function "GetRadiusLayer" that creates the "umbrella" shape. It defines the change in crown radius between the top of the crown and the second ctx.diag.layer above the crown base. Here by default a linear slope
-//! -# crown statistics to be updated, one as input, one as output. These statistics are updated by the function "ModifyCrownStatistic" and then applied across the crown with the function "UpdateCrownStatistic". While they always have to be provided, they do not always have to be used.
-//!
-//! EXAMPLE 1: for adding volume to a voxel field, the input statistic would be +1, for removing volume, -1, the ModifyCrownStatistic function empty, and the update function would simply add the input value to the Voxel3D field
-//!
-//! EXAMPLE 2: for adding leaves to the LAI field, the input statistic would be the tree LAI, or for removing, -LAI, ModifyCrownStatistic would convert the LAI to density within a specific ctx.diag.layer, and the update function would simply add the resulting density values to the ctx.field.LAI3D field
-//! - Input and output variables can be separate types (e.g. int and float) and of different length (e.g. input can be a single variable, output can be a vector. This is needed, for example, to compute PPFD, VPD, Tmp and leafarea in Fluxh)
-//! - In the current implementation, crowns below 3m in crown depth are simply treated as cylinders, this could be changed in future implementations
-template <typename I, typename O, typename M, typename F>
-void LoopLayerUpdateCrownStatistic_template(int row_center, int col_center, float height, float CR, float CD, float fraction_filled_target, int shell_fromtop, float GetRadiusLayer(float, float, float), I CrownStatistic_input, O &CrownStatistic_output, M ModifyCrownStatistic_input, F UpdateCrownStatistic_output)
-{
-    int crown_top = int(height);
-
-    // we start out with 0 actually filled voxels. As a result, the first voxel will always be filled
-    float fraction_filled_actual = 0.0;
-    if (CD <= 3.0)
-    {
-        I CrownStatistic_input_modified;
-        ModifyCrownStatistic_input(CrownStatistic_input, CrownStatistic_input_modified, CD, height, shell_fromtop);
-
-        int crown_intarea_previous = 0;
-        int crown_intarea = GetCrownIntarea(CR);
-        int layer_cylinder = crown_top - shell_fromtop;
-        CircleAreaUpdateCrownStatistic_template(row_center, col_center, crown_intarea_previous, crown_intarea, fraction_filled_target, fraction_filled_actual, layer_cylinder, CrownStatistic_input_modified, CrownStatistic_output, UpdateCrownStatistic_output);
-    }
-    else
-    {
-        // This function computes the extent of the crown at every height ctx.diag.layer, given a specific function
-        // it separates out the innermost sector (a slowly increasing cylinder), and the surrounding parts of the crown
-        // first the metrics with respect to the internal crown structure (i.e. z coordinate with respect to crown base)
-        float crownshell_base = height - CD + 2.0;                   // lower reference point for the crown slope function is two layers up from the crown base
-        float crownshell_extent = height - crownshell_base;          // this is the extent from the "base ctx.diag.layer" to the top
-        float crownshell_extent_toplayer = floor(crownshell_extent); // this is the extent to the lower limit of the toplayer
-        // then we translate the crown coordinates into discretised variables with respect to the absolute location in the voxel field, as needed for location in the voxel field, with layers defined from top to bottom
-        int height_innermost = crown_top - shell_fromtop;
-        int height_toplayer = int(crownshell_base + crownshell_extent_toplayer) - shell_fromtop;
-        int height_baselayer = int(crownshell_base + 1.0) - shell_fromtop;
-
-        // now calculate the two modifications of the input statistic
-        I CrownStatistic_input_innermost;
-        I CrownStatistic_input_outer;
-
-        ModifyCrownStatistic_input(CrownStatistic_input, CrownStatistic_input_innermost, CD, height, shell_fromtop);
-        ModifyCrownStatistic_input(CrownStatistic_input, CrownStatistic_input_outer, CD, crownshell_base, shell_fromtop);
-
-        // now do calculations
-        // first the inner crown shell section that grows dynamically
-        int crown_intarea_previous = 0;
-
-        float radius_innermost = GetRadiusLayer(CR, crownshell_extent, crownshell_extent_toplayer);
-        int crown_intarea_innermost = GetCrownIntarea(radius_innermost);
-        CircleAreaUpdateCrownStatistic_template(row_center, col_center, crown_intarea_previous, crown_intarea_innermost, fraction_filled_target, fraction_filled_actual, height_innermost, CrownStatistic_input_innermost, CrownStatistic_output, UpdateCrownStatistic_output);
-        crown_intarea_previous = crown_intarea_innermost;
-
-        // now loop through the outer crown shell cylinders
-        for (int h_outer = height_toplayer; h_outer >= height_baselayer; h_outer--)
-        {
-            // calculating the radius of the current ctx.diag.layer depending on the respective slopes, to be replaced by function
-            // float radius_height = CR - crown_slope * (h_outer - height_baselayer);    // for the lowest ctx.diag.layer, i.e. h == height_baselayer, radius = t_CR
-            int extent_layerouter = h_outer - height_baselayer;
-            float radius_height = GetRadiusLayer(CR, crownshell_extent, extent_layerouter);
-            int crown_intarea = GetCrownIntarea(radius_height);
-            CircleAreaUpdateCrownStatistic_template(row_center, col_center, crown_intarea_previous, crown_intarea, fraction_filled_target, fraction_filled_actual, h_outer, CrownStatistic_input_outer, CrownStatistic_output, UpdateCrownStatistic_output);
-            crown_intarea_previous = crown_intarea;
-        }
-    }
-}
-
-//! - Global function: the following function is not a member function of the Tree class object, but operates at the tree level and could be converted to member functions. They're therefore defined next to the other tree level functions. Whether conversion to Tree members adds any benefit in terms of performance should be tested.
-//! - this is the template that simulates the actual circling through a crown shell, between a given start and stop position (i.e. between a starting crown area and a stopping crown area)
-//! - If the crown shells are bent, i.e. extend across several canopy layers, the starting position within a lower ctx.diag.layer is the stopping position of the ctx.diag.layer just above
-template <typename I, typename O, typename F>
-void CircleAreaUpdateCrownStatistic_template(int row_center, int col_center, int pos_start, int pos_end, float fraction_filled_target, float &fraction_filled_actual, int height_layer, I CrownStatistic_input, O &CrownStatistic_output, F UpdateCrownStatistic)
-{
-
-    for (int i = pos_start; i < pos_end; i++)
-    {
-        if (fraction_filled_actual > fraction_filled_target)
-        {
-            fraction_filled_actual = (fraction_filled_actual * float(i)) / (float(i) + 1.0);
-        }
-        else
-        {
-            fraction_filled_actual = (fraction_filled_actual * float(i) + 1.0) / (float(i) + 1.0);
-
-            int site_relative = ctx.lookup.LookUp_Crown_site[i];
-            int row, col;
-            row = row_center + site_relative / 51 - 25;
-            col = col_center + site_relative % 51 - 25;
-
-            //            // option for future implementations: choose a direction of the loop (to randomize patterns, etc.)
-            //            // to create less homogeneous and more realistic crowns, this can, for example, be varied between crown layers
-            //            if(direction == 0 || direction == 1){row = row_center + site_relative/51 - 25;}
-            //            else {row = row_center - site_relative/51 + 25;}
-            //            if(direction == 0 || direction == 2){col = col_center + site_relative%51 - 25;}
-            //            else {col = col_center - site_relative%51 + 25;}
-
-            if (row >= 0 && row < ctx.grid.rows && col >= 0 && col < ctx.grid.cols)
-            {
-                int site = col + ctx.grid.cols * row;
-                UpdateCrownStatistic(height_layer, site, CrownStatistic_input, CrownStatistic_output);
-            }
-        }
-    }
-}
-
-// Global function: linear decrease of crown radius
-float GetRadiusSlope(float CR, float crown_extent, float crown_position)
-{
-    float crown_slope = CR * (1.0 - ctx.crown.shape_crown) / crown_extent;
-    float radius = CR - crown_slope * float(crown_position);
-    return (radius);
-}
-
-// Global function: not currently used, but returns the input radius
-float GetRadiusCylinder(float CR, float crown_extent, float crown_position)
-{
-    return (CR);
-}
-
-// Global function: converts floating point crown area into integer value, imposing lower and upper limits
-int GetCrownIntarea(float crown_radius)
-{
-    // crown area
-    float crown_area = PI * crown_radius * crown_radius;
-    int crown_intarea = int(crown_area);      // floor of crown_area to bound area accumulation
-    crown_intarea = max(crown_intarea, 1);    // minimum area of crown (1)
-    crown_intarea = min(crown_intarea, 1963); // maximum area of crown (radius 25), int(3.14*25*25)
-    return (crown_intarea);
-}
-
-// Global function: deduces within-crown densities from LAI with a gradient from 50% in top ctx.diag.layer to 25% in belowtop and 25% in all shells underneath (1 ctx.diag.layer for umbrella-like shape)
-void GetDensitiesGradient(float LAI, float CD, float &dens_top, float &dens_belowtop, float &dens_layer) // RENAMED: ctx.params.dens → dens_layer (was shadowing global ctx.params.dens/ctx.params.ctx.params.dens)
-{
-    if (CD < 2.0)
-    {
-        dens_top = dens_belowtop = dens_layer = LAI / CD;
-    }
-    else if (CD < 3.0)
-    {
-        dens_top = 0.5 * LAI;
-        dens_belowtop = dens_layer = 0.5 * LAI / (CD - 1.0);
-    }
-    else
-    {
-        dens_top = 0.5 * LAI;
-        dens_belowtop = 0.25 * LAI;
-#ifdef CROWN_UMBRELLA
-        dens_layer = 0.25 * LAI;
-#else
-        dens_layer = 0.25 * LAI / (CD - 2.0);
-#endif
-    }
-}
-
-// Global function: deduces within-crown density from LAI, assuming uniform leaf distribution
-void GetDensityUniform(float LAI, float CD, float &dens_layer) // RENAMED: ctx.params.dens → dens_layer
-{
-#ifdef CROWN_UMBRELLA
-    float crownshells_limit = fminf(CD, 3.0);
-    dens_layer = LAI / crownshells_limit;
-#else
-    dens_layer = LAI / CD;
-#endif
-}
-
-// Global function: dummy function when no modification is needed
-void KeepFloatAsIs(float input, float &output, float CD, float height, int layer_fromtop)
-{
-    output = input;
-}
-
-void KeepIntAsIs(int input, int &output, float CD, float height, int layer_fromtop)
-{
-    output = input;
-}
-
-// Global function: a modifying function that converts LAI to the density of a specific ctx.diag.layer, using the GetDensity functions
-//! - modifier for GPP calculation where we need the leaves per ctx.diag.layer to weight our results
-//! - LAI is the input, dens_layer the output
-void LAI2dens_cumulated(float LAI, float &dens_layer, float CD, float height, int layer_fromtop)
-{
-    int crown_top = int(height);
-    int crown_base = int(height - CD);
-    float dens_top, dens_belowtop, dens_lowerlayer; // RENAMED: dens → dens_lowerlayer (base-layer crown density from GetDensitiesGradient; distinct from output param dens_layer)
-#ifdef LAI_gradient
-    GetDensitiesGradient(LAI, CD, dens_top, dens_belowtop, dens_lowerlayer);
-#else
-    GetDensityUniform(LAI, CD, dens_lowerlayer);
-    dens_top = dens_belowtop = dens_lowerlayer;
-#endif
-
-    if (CD < 3.0 && crown_top == crown_base)
-    {
-        dens_layer = LAI; /* full LAI allocation */
-    }
-    else if (CD < 3.0 && (crown_top - layer_fromtop == crown_base))
-    {
-        dens_layer = LAI;
-    }
-    else
-    {
-        float fraction_layer = height - floor(height); /* this is the fraction that each ctx.diag.layer apart from the topmost ctx.diag.layer will extend into the voxel above */
-        if (layer_fromtop == 0)
-            dens_layer = dens_top * fraction_layer;
-        else if (layer_fromtop == 1)
-            dens_layer = dens_top + dens_belowtop * fraction_layer;
-        else if (layer_fromtop == 2)
-            dens_layer = dens_top + dens_belowtop + dens_lowerlayer * fraction_layer;
-        else
-            dens_layer = LAI;
-    }
-}
-
-// Global function: a modifying function that converts LAI to the density of a specific ctx.diag.layer, using the GetDensity functions
-//! - modifier for GPP calculation where we need the leaves per ctx.diag.layer to weight our results
-//! - LAI is the input, dens_layer the output
-void LAI2dens(float LAI, float &dens_layer, float CD, float height, int layer_fromtop)
-{
-
-    int crown_top = int(height);
-    int crown_base = int(height - CD);
-    float dens_top, dens_belowtop, dens_lowerlayer; // RENAMED: dens → dens_lowerlayer (base-layer crown density from GetDensitiesGradient; distinct from output param dens_layer)
-#ifdef LAI_gradient
-    GetDensitiesGradient(LAI, CD, dens_top, dens_belowtop, dens_lowerlayer);
-#else
-    GetDensityUniform(LAI, CD, dens_lowerlayer);
-    dens_top = dens_belowtop = dens_lowerlayer;
-#endif
-
-    if (CD < 3.0 && crown_top == crown_base)
-    {
-        dens_layer = dens_top * CD;
-    }
-    else if (CD < 3.0 && (crown_top - layer_fromtop == crown_base))
-    {
-        float fraction_belowbase = float(crown_base + 1) - (height - CD);
-        dens_layer = dens_lowerlayer * fraction_belowbase;
-    }
-    else
-    {
-        float fraction_layer = height - floor(height);         // this is the fraction that each ctx.diag.layer apart from the topmost ctx.diag.layer will extend into the voxel above
-        float fraction_layer_fromabove = 1.0 - fraction_layer; // the inverse of the fraction above
-
-        if (layer_fromtop == 0)
-            dens_layer = dens_top * fraction_layer;
-        else if (layer_fromtop == 1)
-            dens_layer = dens_top * fraction_layer_fromabove + dens_belowtop * fraction_layer;
-        else if (layer_fromtop == 2)
-            dens_layer = dens_belowtop * fraction_layer_fromabove + dens_lowerlayer * fraction_layer;
-        else
-            dens_layer = dens_lowerlayer * fraction_layer_fromabove;
-    }
-}
-
-// Global function: update of ctx.field.LAI3D field, called by CalcLAI()
-void UpdateLAI3D(int height, int site, float dens_layer, float &LA_cumulated) // RENAMED: ctx.params.dens → dens_layer
-{
-    ctx.field.LAI3D[height][site + ctx.grid.SBORD] += dens_layer;
-    LA_cumulated += dens_layer;
-}
-
-// Global function: remove outliers in canopy height model (CHM); vector option
-#ifdef CHM_SPIKEFREE
-void UpdateCHMvector(int height, int site, float noinput, vector<int> &chm)
-{
-    if (chm[site] < height)
-        chm[site] = height;
-}
-// Global function: remove outliers in canopy height model (CHM)
-void UpdateCHM(int height, int site, float noinput, int *chm)
-{
-    if (chm[site] < height)
-        chm[site] = height;
-}
-#endif
-
-void OutputCrownSliced(int height, int site, int row_slice, vector<float> &output_statistics)
-{
-    int row_current = site / ctx.grid.cols;
-    int col_current = site % ctx.grid.cols;
-    if (row_current == row_slice && col_current >= ctx.crown.mincol_visual && col_current < ctx.crown.maxcol_visual)
-    {
-        output_visual[1] << ctx.time.iter << "\t" << row_current << "\t" << col_current << "\t" << height;
-        for (int i = 0; i < output_statistics.size(); i++)
-        {
-            output_visual[1] << "\t" << output_statistics[i];
-        } // we only output tree parts that fall in the current slice extent
-        output_visual[1] << endl;
-    }
-};
-
-// Global function: PPFD retrieval for function leafarea_max()
-void GetPPFDabove(int height, int site, float noinput, float (&ppfd_CA)[2])
-{
-    // First get voxel field densities
-    float absorb_prev = ctx.field.LAI3D[height + 1][site + ctx.grid.SBORD];
-    int intabsorb = CalcIntabsorb(absorb_prev);
-
-    // Obtain PPFD for the voxel, and also record the circled area
-    ppfd_CA[0] += ctx.climate.WDailyMean * ctx.lookup.LookUp_flux[intabsorb];
-    ppfd_CA[1] += 1.0; // add area
-}
-
-// Global function: calculates the canopy environment
-//! - this function adds to the environmental variables provided in canopy_environment_cumulated
-//! - the PPFD, VPD, Tmp and leafarea_layer retrieval function for Fluxh()
-#ifdef WATER
-void GetCanopyEnvironment(int height, int site, float dens_layer, float (&canopy_environment_cumulated)[6]) // RENAMED: ctx.params.dens → dens_layer
-{
-#else
-void GetCanopyEnvironment(int height, int site, float dens_layer, float (&canopy_environment_cumulated)[4]) // RENAMED: ctx.params.dens → dens_layer
-{
-#endif
-    // first get voxel field densities
-    float absorb_prev = ctx.field.LAI3D[height + 1][site + ctx.grid.SBORD];
-    float absorb_curr = ctx.field.LAI3D[height][site + ctx.grid.SBORD];
-    float absorb_delta = absorb_curr - absorb_prev;
-    if (absorb_delta < 0.0)
-        absorb_delta = 0.0; // eliminate rounding errors
-    int intabsorb = CalcIntabsorb(absorb_prev, absorb_delta);
-#ifdef WATER
-    int intincident = CalcIntabsorb(absorb_prev);
-#endif
-
-#ifdef FULL_CLIMATE
-    // Obtain PPFD, VPD and T for the voxel
-    float PPFD_voxel = ctx.lookup.LookUp_flux_absorption[intabsorb];
-    float VPD_voxel = ctx.lookup.LookUp_VPD[intabsorb];
-    float T_voxel = ctx.lookup.LookUp_T[intabsorb];
-#ifdef WATER
-    float PPFD_voxel_incident = ctx.lookup.LookUp_flux[intincident];
-    float ExtinctLW_voxel = ctx.lookup.LookUp_ExtinctLW[intincident];
-#endif // WATER
-
-#else // FULL_CLIMATE
-      // Obtain PPFD, VPD and T for the voxel
-    float PPFD_voxel = ctx.climate.WDailyMean * ctx.lookup.LookUp_flux_absorption[intabsorb];
-    float VPD_voxel = ctx.climate.VPDDailyMean * ctx.lookup.LookUp_VPD[intabsorb];
-    float T_voxel = ctx.climate.tDailyMean - ctx.lookup.LookUp_T[intabsorb];
-#ifdef WATER
-    float PPFD_voxel_incident = ctx.climate.WDailyMean * ctx.lookup.LookUp_flux[intincident];
-    float ExtinctLW_voxel = ctx.lookup.LookUp_ExtinctLW[intincident];
-#endif // WATER
-
-#endif // FULL_CLIMATE
-
-    // Add the three variables up, weighted by leaf density inside voxel
-    canopy_environment_cumulated[0] += dens_layer;
-    canopy_environment_cumulated[1] += PPFD_voxel * dens_layer;
-    canopy_environment_cumulated[2] += VPD_voxel * dens_layer;
-    canopy_environment_cumulated[3] += T_voxel * dens_layer;
-#ifdef WATER
-    canopy_environment_cumulated[4] += PPFD_voxel_incident * dens_layer;
-    canopy_environment_cumulated[5] += ExtinctLW_voxel * dens_layer;
-#endif
-
-    // if (canopy_environment_cumulated[1]<=0 || dens_layer < 0.05 || absorb_delta < 0.05) {
-    //  cout << "Warning in GetCanopyEnvironment, PPFD <=0; PPFD_voxel=" << PPFD_voxel << "; dens_layer=" << dens_layer << endl;
-    //  cout << "ctx.climate.WDailyMean=" << ctx.climate.WDailyMean << "; ctx.lookup.LookUp_flux_absorption[intabsorb]=" << ctx.lookup.LookUp_flux_absorption[intabsorb] << "; intabsorb=" << intabsorb << "; absorb_prev=" << absorb_prev << "; absorb_delta=" << absorb_delta << endl;
-    //  }
-}
-
-// Global function: calculates packing densities
-void AddCrownVolumeLayer(int row_center, int col_center, float height, float CR, float CD, int crownvolume[70])
-{
-    int crown_top = int(height);
-    int crown_base = int(height - CD);
-    if (CD <= 3.0)
-    {
-        // for the smallest crowns it is simply cylinder rings being filled up
-        int crown_intarea = GetCrownIntarea(CR);
-        for (int h = crown_top; h >= crown_base; h--)
-            crownvolume[h] += crown_intarea;
-    }
-    else
-    {
-        // For the rest of the crown, we go through different crown shells. We separate out the innermost sector (a slowly increasing cylinder), and the surrounding parts of the crown
-        // first the metrics with respect to the internal crown structure (i.e. z coordinate with respect to crown base)
-        float crownshell_base = height - CD + 2.0;                   // lower reference point for the crown slope function is two layers up from the crown base
-        float crownshell_extent = height - crownshell_base;          // this is the extent from the "base ctx.diag.layer" to the top
-        float crownshell_extent_toplayer = floor(crownshell_extent); // this is the extent to the lower limit of the toplayer
-        // then we translate the crown coordinates into discretised variables with respect to the absolute location in the voxel field, as needed for location in the voxel field, with layers defined from top to bottom
-        int shell_fromtop = 0;
-        int height_innermost = crown_top - shell_fromtop;
-        int height_toplayer = int(crownshell_base + crownshell_extent_toplayer) - shell_fromtop;
-        int height_baselayer = int(crownshell_base + 1.0) - shell_fromtop;
-
-        // now do calculations
-        // first the inner crown shell section that grows dynamically
-        float radius_innermost = GetRadiusSlope(CR, crownshell_extent, crownshell_extent_toplayer);
-        int crown_intarea_innermost = GetCrownIntarea(radius_innermost);
-        for (int h = height_innermost; h >= crown_base; h--)
-        {
-            crownvolume[h] += crown_intarea_innermost;
-        }
-        // now loop through the outer crown shell cylinders
-        for (int h_outer = height_toplayer; h_outer >= crown_base; h_outer--)
-        {
-            // calculating the radius of the current ctx.diag.layer depending on the respective slopes, to be replaced by function
-            // float radius_height = CR - crown_slope * (h_outer - height_baselayer);    // for the lowest ctx.diag.layer, i.e. h == height_baselayer, radius = t_CR
-            int extent_layerouter = max(h_outer - height_baselayer, 0); // we also fill up underneath the baselayer
-            float radius_height = GetRadiusSlope(CR, crownshell_extent, extent_layerouter);
-            int crown_intarea = GetCrownIntarea(radius_height);
-
-            crownvolume[h_outer] += (crown_intarea - crown_intarea_innermost);
-        }
-    }
-}
-#endif
 
 // ###########################################
 // ###########################################
